[Agent] add GameStateSerializer utility
